Move timeslots into /campuses and sort alphabetically
