+++ package.json
@@ -55,25 +55,25 @@
     "lodash.set": "^4.3.2",
     "lodash.transform": "^4.6.0",
     "lodash.xor": "^4.5.0",
-    "luxon": "^1.22.0",
-    "mongoose": "^5.9.2",
+    "luxon": "^1.24.1",
+    "mongoose": "^5.9.13",
     "nanoid": "^2.1.11",
     "nconf": "^0.10.0",
     "nodemailer": "^6.4.3",
     "pino": "^6.2.0",
     "prom-client": "^12.0.0",
-    "qs": "^6.9.1",
+    "qs": "^6.9.4",
     "socket.io": "^2.3.0",
     "socket.io-redis": "^5.2.0",
     "sprintf-lite": "^1.3.0",
     "timezone-validator": "^1.0.3",
-    "twilio": "^3.39.5",
+    "twilio": "^3.43.0",
     "validator": "^12.2.0",
-    "web-push": "^3.4.3"
+    "web-push": "^3.4.4"
   },
   "devDependencies": {
     "@std/esm": "^0.26.0",
-    "acorn": "^7.1.0",
+    "acorn": "^7.2.0",
     "babel-eslint": "^10.0.3",
     "chai": "^4.2.0",
     "eslint": "^6.8.0",
