+++ routes/campuses.mjs
@@ -1,11 +1,12 @@
 import mask from 'json-mask';
 import generateCRUD from '../helpers/abstract-route';
 import Campus from '../models/campus';
-import usersRoutes from './campuses/users';
-import driversRoutes from './campuses/drivers';
-import driversPositionsRoutes from './campuses/drivers-positions';
 import carsRoutes from './campuses/cars';
+import driversPositionsRoutes from './campuses/drivers-positions';
+import driversRoutes from './campuses/drivers';
 import statsRoutes from './campuses/stats';
+import timeSlots from './campuses/time-slots';
+import usersRoutes from './campuses/users';
 import {
   CAN_CREATE_CAMPUS,
   CAN_EDIT_CAMPUS, CAN_EDIT_SELF_CAMPUS,
@@ -59,15 +60,17 @@ const router = generateCRUD(Campus, {
   },
 });
 
-router.use('/:campus_id/users', usersRoutes.routes());
-router.use('/:campus_id/users', usersRoutes.allowedMethods());
-router.use('/:campus_id/drivers', driversRoutes.routes());
-router.use('/:campus_id/drivers', driversRoutes.allowedMethods());
-router.use('/:campus_id/drivers-positions', driversPositionsRoutes.routes());
-router.use('/:campus_id/drivers-positions', driversPositionsRoutes.allowedMethods());
-router.use('/:campus_id/cars', carsRoutes.routes());
 router.use('/:campus_id/cars', carsRoutes.allowedMethods());
-router.use('/:campus_id/stats', statsRoutes.routes());
+router.use('/:campus_id/cars', carsRoutes.routes());
+router.use('/:campus_id/drivers', driversRoutes.allowedMethods());
+router.use('/:campus_id/drivers', driversRoutes.routes());
+router.use('/:campus_id/drivers-positions', driversPositionsRoutes.allowedMethods());
+router.use('/:campus_id/drivers-positions', driversPositionsRoutes.routes());
 router.use('/:campus_id/stats', statsRoutes.allowedMethods());
+router.use('/:campus_id/stats', statsRoutes.routes());
+router.use('/:campus_id/time-slots', timeSlots.allowedMethods());
+router.use('/:campus_id/time-slots', timeSlots.routes());
+router.use('/:campus_id/users', usersRoutes.allowedMethods());
+router.use('/:campus_id/users', usersRoutes.routes());
 
 export default router;
